requirements.txt
argon2-cffi
sqlalchemy
psycopg2